requirements.txt
asgiref==3.9.1
Django==5.2.5
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-environ==0.12.0
pytest-django==4.11.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
sqlparse==0.5.3
python-dotenv==1.1.1
pytest==8.4.1
pyflakes==3.4.0